Folder actor moved to commons module
