+++ publisher-commons/src/main/java/nl/idgis/publisher/folder/ChannelCursor.java
@@ -1,13 +1,13 @@
-package nl.idgis.publisher.provider.folder;
+package nl.idgis.publisher.folder;
 
 import java.nio.ByteBuffer;
 import java.nio.channels.AsynchronousFileChannel;
 import java.nio.channels.CompletionHandler;
 import java.util.concurrent.CompletableFuture;
 
+import nl.idgis.publisher.folder.messages.Eof;
+import nl.idgis.publisher.folder.messages.FileChunk;
 import nl.idgis.publisher.protocol.messages.Failure;
-import nl.idgis.publisher.provider.folder.messages.Eof;
-import nl.idgis.publisher.provider.protocol.FileChunk;
 import nl.idgis.publisher.stream.StreamCursor;
 
 import akka.actor.ActorRef;
